refactor: update entity builders and directors
