+++ src/components/users/entity/userBuilder.ts
@@ -1,4 +1,5 @@
 import bcrypt from 'bcrypt';
+
 import { User } from '.';
 
 interface IUserBuilder {
