+++ src/components/wages/entity/monthlyWagesBuilder.ts
@@ -1,14 +1,43 @@
 import { Wage } from '.';
-import { WageBuilder } from './wagesDirector';
 
-export class WageDirector {
-  constructor(private builder: WageBuilder) {}
+export class WageBuilder {
+  private user_id!: string;
+  private amount: number = 0.0;
+  private date: Date = new Date();
+  private exchange_rate: number = 0.0;
+  private created_at: Date = new Date();
+  private updated_at: Date = new Date();
+  private deleted_at?: Date;
 
-  construct(user_id: string, amount: number, date: Date): Wage {
-    return this.builder
-      .setUserId(user_id)
-      .setAmount(amount)
-      .setMonth(date)
-      .build();
+  setUserId(user_id: string): this {
+    this.user_id = user_id;
+    return this;
+  }
+
+  setAmount(amount: number): this {
+    this.amount = amount;
+    return this;
+  }
+
+  setMonth(date: Date): this {
+    this.date = date;
+    return this;
+  }
+
+  setExchangeRate(exchange_rate: number): this {
+    this.exchange_rate = exchange_rate;
+    return this;
+  }
+
+  build(): Wage {
+    return new Wage(
+      this.user_id,
+      this.amount,
+      this.date,
+      this.exchange_rate,
+      this.created_at,
+      this.updated_at,
+      this.deleted_at
+    );
   }
 }
